Initial public upload
This is the main upload of the zooplankton trait database and the scripts in reproducing the figures in Pata & Hunt (for submission) manuscript. This also contains a template for how to subset trait information from the database and methods in estimating missing trait values.

Note that the figures folder is included in the gitignore.